+++ test/correctness/PySys_internal_184/Input/NestedNotVerified/run.py
@@ -1,3 +1,4 @@
+# -*- coding: latin-1 -*-
 
 from pysys.constants import *
 from pysys.basetest import BaseTest